Move DiscordSender, RandomInsult, RandomQuote, RandomTaco and WhoStoleTheCookie
