+++ Program.cs
@@ -191,131 +191,6 @@ namespace UtilityBelt
 
     #region Choice Processors
 
-    #region Discord Sender
-
-    private static void DiscordWebhook(IOptions<SecretsModel> options)
-    {
-      logger.LogInformation($"User Choice : {nameof(DiscordWebhook)}");
-      string whook = options.Value.DiscordWebhook;
-
-      if (String.IsNullOrEmpty(whook))
-      {
-        Console.WriteLine("Whoops! You dont have a webhook defined in your config!"); return;
-      }
-
-      Console.Write("Enter the message:");
-      string msg = Console.ReadLine();
-
-      WebHookContent cont = new WebHookContent()
-      {
-        content = msg
-      };
-      string json = JsonSerializer.Serialize(cont);
-
-      using (var www = new HttpClient())
-      {
-        var content = new StringContent(json);
-        content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
-        var task = www.PostAsync(whook, content);
-        task.Wait();
-      }
-      Console.WriteLine("Message sent!");
-    }
-
-    #endregion Discord Sender
-
-    #region Random quote
-
-    private static void RandomQuote()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomQuote)}");
-      string content = string.Empty;
-      string quoteUrl = "https://api.forismatic.com/api/1.0/?method=getQuote&lang=en&format=json";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(quoteUrl);
-      }
-      QuoteModel quote = JsonSerializer.Deserialize<QuoteModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(quote.QuoteText);
-      Console.WriteLine($"--{quote.QuoteAuthor}");
-      Console.WriteLine();
-    }
-
-    #endregion Random quote
-
-    #region Random insult
-
-    private static void RandomInsult()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomInsult)}");
-      string content = string.Empty;
-      string apiUrl = "https://evilinsult.com/generate_insult.php?lang=en&type=json";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(apiUrl);
-      }
-      EvilInsultModel insultResponse = JsonSerializer.Deserialize<EvilInsultModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(HttpUtility.HtmlDecode(insultResponse.Insult));
-      Console.WriteLine();
-    }
-
-    #endregion Random insult
-
-    #region Who stole the cookie
-
-    private static void CookieAccusation()
-    {
-      logger.LogInformation($"User Choice : {nameof(CookieAccusation)}");
-      string content = string.Empty;
-      string suspectUrl = "https://randomuser.me/api/?inc=name&format=json";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(suspectUrl);
-      }
-      CookieSuspectModel cookieSuspect = JsonSerializer.Deserialize<CookieSuspectModel>(content);
-      string suspectFullName = cookieSuspect.Results[0].Name.First + " " + cookieSuspect.Results[0].Name.Last;
-      Console.WriteLine();
-      Console.ForegroundColor = ConsoleColor.Yellow;
-      Console.WriteLine("");
-      Console.WriteLine("Jacques Clouseau: " + suspectFullName + " stole the cookie from the cookie jar.");
-      Console.WriteLine("");
-      Console.WriteLine(suspectFullName + ": Who, me?");
-      Console.WriteLine("");
-      Console.WriteLine("Jacques Clouseau: Yes, you!");
-      Console.WriteLine("");
-      Console.WriteLine(suspectFullName + ": Couldn't be!");
-      Console.WriteLine("");
-      Console.WriteLine("Jacques Clouseau: Then who?");
-      Console.WriteLine();
-    }
-
-    #endregion Who stole the cookie
-
-    #region Random taco recipe
-
-    private static void RandomTaco()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomTaco)}");
-      string content = string.Empty;
-      string tacoRandomizerUrl = "http://taco-randomizer.herokuapp.com/random/";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(tacoRandomizerUrl);
-      }
-
-      TacoRecipeModel recipe = JsonSerializer.Deserialize<TacoRecipeModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"Why not try {recipe.BaseLayer.Name} with {recipe.Mixin.Name},");
-      Console.WriteLine(@$"seasoned with {recipe.Seasoning.Name}, topped off with {recipe.Condiment.Name}");
-      Console.WriteLine(@$"and wrapped in delicious {recipe.Shell.Name}.");
-
-      Console.WriteLine();
-    }
-
-    #endregion Random taco recipe
-
     #region Covid-19
 
     private static void Covid19()
@@ -842,11 +717,6 @@ namespace UtilityBelt
     #endregion Choice Processors
 
     #region Utility
-
-    internal class WebHookContent
-    {
-      public string content { get; set; }
-    }
 
     private enum BooleanAliases
     {
